Add insert macro option to default task settings panel
