+++ src/main/kotlin/com/elpassion/mainframerplugin/action/configure/configurator/ConfiguratorDialog.kt
@@ -2,16 +2,11 @@ package com.elpassion.mainframerplugin.action.configure.configurator
 
 import com.elpassion.mainframerplugin.action.configure.configurator.ui.ConfiguratorForm
 import com.elpassion.mainframerplugin.common.*
-import com.intellij.ide.macro.MacrosDialog
+import com.elpassion.mainframerplugin.common.ui.InsertMacroActionListener
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.wm.IdeFocusManager
 import com.intellij.ui.CollectionComboBoxModel
 import io.reactivex.Maybe
-import java.awt.event.ActionEvent
-import java.awt.event.ActionListener
 import javax.swing.JComponent
-import javax.swing.JTextField
-import javax.swing.text.BadLocationException
 
 class ConfiguratorDialog(project: Project,
                          val defaultValues: ConfiguratorIn,
@@ -57,22 +52,3 @@ private fun showConfigurationDialog(project: Project, defaultValues: Configurato
                 emitter.onComplete()
             }).show()
         }
-
-private class InsertMacroActionListener(private val myTextField: JTextField,
-                                        private val project: Project) : ActionListener {
-
-    override fun actionPerformed(e: ActionEvent) {
-        val dialog = MacrosDialog(project)
-        if (dialog.showAndGet() && dialog.selectedMacro != null) {
-            val macro = dialog.selectedMacro.name
-            val position = myTextField.caretPosition
-            try {
-                myTextField.document.insertString(position, "$$macro$", null)
-                myTextField.caretPosition = position + macro.length + 2
-            } catch (ignored: BadLocationException) {
-            }
-
-        }
-        IdeFocusManager.findInstance().requestFocus(myTextField, true)
-    }
-}
