+++ src/main/kotlin/com/elpassion/mainframerplugin/task/ui/TaskDefaultSettingsPanel.java
@@ -1,6 +1,7 @@
 package com.elpassion.mainframerplugin.task.ui;
 
 import com.elpassion.mainframerplugin.common.*;
+import com.elpassion.mainframerplugin.common.ui.InsertMacroActionListener;
 import com.elpassion.mainframerplugin.task.MainframerTaskDefaultSettingsProvider;
 import com.elpassion.mainframerplugin.task.TaskData;
 import com.intellij.openapi.options.ConfigurationException;
@@ -22,6 +23,7 @@ public class TaskDefaultSettingsPanel {
     public JPanel panel;
     private LabeledComponent<TextFieldWithBrowseButton> mainframerToolHolder;
     public JTextField remoteMachineField;
+    private JButton insertMacroButton;
     private ToolConfiguration toolConfiguration;
 
     public TaskDefaultSettingsPanel(Project project, MainframerTaskDefaultSettingsProvider settingsProvider, ToolConfiguration toolConfiguration) {
@@ -37,6 +39,7 @@ public class TaskDefaultSettingsPanel {
         mainframerToolField.addBrowseFolderListener(textBrowseFolderListener);
         mainframerToolHolder = new LabeledComponent<>();
         mainframerToolHolder.setComponent(mainframerToolField);
+        insertMacroButton.addActionListener(new InsertMacroActionListener(buildCommandField, project));
     }
 
     public Boolean isModified() {
